requirements.txt
# Core API
fastapi==0.110.0
uvicorn[standard]==0.27.0
twilio==9.0.0
python-multipart==0.0.9

# Environment
python-dotenv==1.0.1

# AI & LLM
groq>=0.37.0
langchain>=1.2.0
langchain-community>=0.4.0
langchain-text-splitters>=1.1.0
langgraph>=1.1.0
langchain-groq>=1.1.0
httpx>=0.27,<0.28              # groq SDK uses deprecated 'proxies' kwarg removed in 0.28

# Embeddings & ML
sentence-transformers==2.5.1
torch==2.2.0

# Vector Database
qdrant-client==1.8.0

# Web Scraping
beautifulsoup4==4.12.3
html2text==2024.2.26
requests==2.31.0
selenium==4.18.0
duckduckgo-search==5.1.0

# Data Processing
pandas==2.2.1
tiktoken==0.6.0

# Validation
pydantic==2.6.3

# Image & Document Processing
Pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
pypdf==4.0.2
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# UI
streamlit==1.32.0
audio-recorder-streamlit==0.0.10
